% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm_data.R
\name{ddm_data}
\alias{ddm_data}
\title{Simulate DDM Data}
\usage{
ddm_data(
  n = 100,
  drift = 0,
  bs = 1,
  bias = 0.5,
  ndt = 0.2,
  ndt_var = 0.1,
  include_parameters = FALSE,
  ...
)
}
\arguments{
\item{n}{Number of observations.}

\item{drift}{Drift parameter. Typical range: [-5, 5].}

\item{bs}{boundary separation parameter. Typical range: [0.5, 2].}

\item{bias}{Bias parameter. Default to 0.5*bs.}

\item{ndt}{Non-decision time parameter.  Typical range: [0.1, 0.5] s.}

\item{ndt_var}{Variability of non-decision time. Typical range: [0, 0.2] s.}

\item{include_parameters}{Whether to add the parameters to the data.}

\item{...}{Other arguments to pass to other functions, such as additional controls to \code{\link{ddm_traces}()}.}
}
\description{
Simulate response time data with specific parameters following a diffusion model.
The data are generated using \code{\link[rtdists]{rdiffusion}()}, so that check its documentation.
}
\examples{
df <- ddm_data(n = c(100, 100), drift = c(1, 0))
summary(df$data)

}
