% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm_plot.R
\name{ddm_plot}
\alias{ddm_plot}
\alias{ddm_plot_upper}
\alias{ddm_plot_lower}
\alias{ddm_plot_traces}
\title{Plot DDM Data}
\usage{
ddm_plot(data, ...)

ddm_plot_upper(data, density = NULL, xlim = c(0, 2), theme = "default", ...)

ddm_plot_lower(data, density = NULL, xlim = c(0, 2), theme = "default", ...)

ddm_plot_traces(
  traces,
  alpha = 0.8,
  linewidth = 0.5,
  xlim = c(0, 2),
  theme = "default",
  ...
)
}
\arguments{
\item{data}{Either object from \code{\link{ddm_data}()}) or dataframe containing
columns "rt", "response" (lower or upper) and "condition".}

\item{...}{Other arguments to pass to other functions, such as additional controls to \code{\link[graphics]{hist}()}.}

\item{density}{Data for density lines (typically obtained from \code{\link{ddm_data}()}).}

\item{xlim, linewidth, alpha}{Aesthetic parameters.}

\item{theme}{Can be \code{NULL} (useful for customizing) or "default".}

\item{traces}{Data for trace lines (typically obtained from \code{\link{ddm_data}()}).}
}
\description{
Plot data.
}
\examples{
library(ggplot2)
library(patchwork)

# Simulate data
sim <- ddm_data(n=1000, drift = c(1, 0), n_traces = 20)

ddm_plot(sim, xlim = c(0, 1))

# Customize
data <- sim$data
density <- sim$density
traces <- sim$traces

ddm_plot_upper(data, density = density, theme = NULL) /
  ddm_plot_traces(traces, theme = NULL) /
  ddm_plot_lower(data, density = density, theme = NULL)

}
