% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm_traces.R
\name{ddm_traces}
\alias{ddm_traces}
\title{Plot DDM Data}
\usage{
ddm_traces(
  n_traces = 5,
  time_max = 2,
  step = 0.01,
  drift = 0,
  bs = 1,
  bias = 0.5,
  ndt = 0.2,
  ndt_var = 0.1,
  ...
)
}
\arguments{
\item{n_traces}{Number of traces.}

\item{time_max}{Span of the x-axis in seconds.}

\item{step}{Step size in seconds.}

\item{drift}{Drift parameter. Typical range: [-5, 5].}

\item{bs}{boundary separation parameter. Typical range: [0.5, 2].}

\item{bias}{Bias parameter. Default to 0.5*bs.}

\item{ndt}{Non-decision time parameter.  Typical range: [0.1, 0.5] s.}

\item{ndt_var}{Variability of non-decision time. Typical range: [0, 0.2] s.}

\item{...}{Not used for now.}
}
\description{
Plot data.
}
\examples{

data_traces <- ddm_traces(n_traces = 5)
data_traces

}
